separate the tool functions of HTML
